refactor(frontend, services): create services folder
