--- apps/frontend/src/lib/api/endpoints/index.ts
@@ -1,1 +1,0 @@
-export * from "./globals"
