+++ apps/frontend/src/services/cms/faq/FaqService.test.ts
@@ -2,7 +2,7 @@ import { GetFaqResponseSchema } from "@repo/shared"
 import type { Faq } from "@repo/shared/payload-types"
 import { createSimpleSharedFAQItem } from "@repo/ui/test-config/mocks/FAQ.mock"
 import { describe, expect, it, vi } from "vitest"
-import { getFaq } from "./globals"
+import { getFaq } from "./FaqService"
 
 const mockFaq: Faq = {
   id: "1",
